finished Robot interface, still need to work on GUI
GUI has some stuff to work on, ex: msking a feilf with collision detection, and/or adding more things to the oversll appearence. This commit is Stable (at least using a macbook using VS Code 2019). there is extra run code in Robot.java, take it out if you can compile it using WPILIB.